Adjust some gen-gc array sizes.
Tuning some parameters by experimenting with
timings.

* gc.c (CHECKOBJ_VEC_SIZE): Reduce to 2*HEAP_SIZE;
this doesn't have to be equal in size to FRESHOBJ_VEC_SIZE;
entries into the checkobj array are not expected to
be as frequent as in the freshobj array.
(MUTOBJ_VEC_SIZE): On the other hand, at HEAP_SIZE/4,
this may be too small. Increasing to 2*HEAP_SIZE.
Overall, the storage for these arrays is reduced by
over 25%, with seemingly better performance.