bugfix: quasilit read/print consistency, part 2.
In this patch commit I'm addressing the issue introduced in
part 1 that expressions in @(output) blocks are still using
(sys:expr ...) wrapping, but are passed down to an evaluator
which now expects unwrapped expressions now.

As part of this change, I'm changing the representation of
@expr from (sys:expr . expr) to (sys:expr expr).

* eval.c (format_field): Adjust access to sys:expr
expression based on new representation.
(transform_op): Likewise.

* lib.c (obj_print_impl): Likewise.

* match.c (dest_bind): Likewise.
(do_txeval): Likewise.
(do_output_line): Likewise, in some compat code.  Here is the
fix for the issue: when calling tx_subst_vars, we pass a list
of one element containing the expression, not wrapped in
sys:expr. Previously, we passed a one-element list containing
the sys:expr.

* parser.y (o_elem): If a list occurs in the syntax, represent
it as (sys:expr list) rather than (sys:expr . list).
(list): Do the same for @ n_expr syntax.
(expand_meta, make_expr): Harmonize with the representation
change.