Don't throw when printing (sys:quasi . atom).
We do not have perfect read/print consistency for
quasiliterals. Programs can construct quasiliterals which
cause the printer to throw exceptions, or which don't
print such that the same object is read back.

However, at least we can handle some trivial cases.

In particular, the object (sys:quasi . #<function>)
occurs in  the system, as the top-level binding of
the quasi operator. With this change, we can print that
instead of throwing.

* lib.c (obj_print_impl): Check that a quasiliteral or
quasi-list-literal is a list with at least one argument.
Don't print (sys:quasi) or (sys:quasi . non-nil-atom) in the
notation.