conformance: handle macro wrong clause syntax.
Contrary to the documentation, handle doesn't in fact have the
same syntax as catch. It passes the exception symbol to
clauses as the leftmost argument, followed by the exception
arguments, whereas catch passes only the exception arguments.

* share/txr/stdlib/except.tl (handle): Do not pass the
exception sybmol as the leftmost argument, unless operating
in TXR 161 or earlier compatibility.

* tests/012/except.tl: drop exception symbol argument from
handle clause.

* txr.1: Compatibility note added.