bugfix: previous loop elision fix in awk.
* share/txr/stdlib/awk.tl (awk): We must generate the
code to bind the awk-fun under the same conditions that
control whether the loop call is generated to which it
is an argument. Otherwise we end up with an unbound var
reference: the loop isn't elided because begin/end
actions are present, but the call refers to a nonexistent
variable.