bugfix: , *sym printed as ,*sym.
We are lacking read/print consistency in the handling of
unquotes applied to symbols whose names begin with a star.

* lib.c (unquote_star_check): New static function.
(obj_print_impl): Use unquote_star check when printing
an unquote to determine whether a space is needed so
that the result doesn't read back as a ,* splice.

* txr.1: Change "should" to "must": the whitespace is
absolutely required in , *x*. Adding more discussion
as a dialect note.