Cancel in-progress gc if handling async signal.
If a signal goes off in the middle of gc that is
routed to an async lambda handler, we should cancel
gc, so the handler finds the heap in a good state.
Of course, the handler must not return in this
case, or else the interrupted gc will try to resume.

* gc.c (inprogress): New static flag.
(gc): Increment and decrement inprogress variable.
Abort if re-entered.
(gc_inprogress): New function.
(unmark): Do not clear FREE flags, only REACHABLE.
(gc_cancel): New function.

* gc.h (gc_inprogress, gc_cancel): Declared.

* signal.c (sig_handler): In the async lambda
case, check for an in-progress gc and cancel it.