bugfix: :whole and :form allow destructuring.
In bind_macro_params allow the argument of :whole and :form
to be a destructuring pattern. Yet, in expand_params, we
are wrongly requiring it to be a bindable symbol, thus
blocking the feature.

* eval.c (expand_opt_params_rec): Only require the argument
of a keyword to be a bindable symbol if that keyword is :env.
(expand_params_rec): Some refactoring here not to recursively
expand keyword as parameters, removing a hacky test in the
atom case. Also car_ex is renamed to param_ex. Other than
that, same fix as in expand_opt_params_rec.