Add useful sys:bits function.
sys:bits converts a Lisp value to an integer whose value is
the object's bit pattern interpreted as a pure binary number.
(Only the "unboxed" part of the object that is stored in
variables or passed into functions, not any "boxded" heap part
which is referenced.)

this holds:

  (eq a b)  <-->  (= (sys:bits a) (sys:bits b))

Two values a and b are the same object iff their sys:bits
values are the same integer.

* arith.c (bits): New static function.
(arith_init): Register bits as sys:bits.