New system function: expand-with-free-refs.
This interface to the expander returns not only the expanded
form, but also a list of the free variables and functions
occurring in that form.

This interface to the expander works by installing a handler
which intercepts and muffles warnings.  When a warning occurs
indicating an unbound variable or function, the information is
retained.  The expander then returns the information along
with the expanded form.

* eval.c (gather_free_refs): New static function.
(expand_with_free_refs): New function.
(eval_init): Register sys:expand-with-free-refs intrinsic.