awk macro: warn about invalid function use in rng.
Improving the new feature to handle functions also,
and improve the diagnostic by reporting the specific
to or from subform of which contains the reference.

* share/txr/stdlib/awk.tl (sys:awk-code-move-check): Take a
an argument indicating the kinds of bindings being checked,
symbols or functions, and an extra form argument.
(sys:awk-mac-let): Pass new arguments to
sys:awk-code-move-check and also call it two more times
to report on functions also.