bugfix: :counter var not registered tentative.
This test case suffers a spurious unbound variable warning:

  @(collect :counter x)
  @(bind y @x)
  @(end)

* match.c (match_expand_keyword_args): Register the :counter
variable as a tentative definition with match_reg_var.
We don't do this for :vars because they are not newly
introduced: the :vars construct doesn't bind. The :counter
feature is the only keyword feature in collect which binds
a new variable.